Added in some basic CSS and general idea for Pokedex App
The idea is that once a pokemon's name is typed in, it will spawn that pokemon's sprite (ideally a 3d or at least gen 5 one). in the middle of the page while it's information is placed on the right.

Also plan on adding arrows on both sides of the avatar to cycle between each pokemon by pokedex #.

A list of all pokemon will be available as well perhaps on another page or maybe on the bottom before the footer.

Also plan on adding a random button that brings you to a random pokemon.

Not sure how any of these will work so far but I'm hoping to learn them in the upcoming chapters!